+++ lib/game/game.dart
@@ -1,16 +1,17 @@
-import 'dart:ui';
 import 'dart:async';
+import 'dart:ui';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:flutter/widgets.dart' hide WidgetBuilder;
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter/widgets.dart' hide WidgetBuilder;
 
+import '../assets/images.dart';
+import '../extensions/vector2.dart';
+import '../keyboard.dart';
 import 'widget_builder.dart';
 
-import '../keyboard.dart';
-import '../assets/images.dart';
 
 /// Represents a generic game.
 ///
@@ -38,7 +39,7 @@ abstract class Game {
   /// This is the resize hook; every time the game widget is resized, this hook is called.
   ///
   /// The default implementation does nothing; override to use the hook.
-  void resize(Size size) {}
+  void resize(Vector2 size) {}
 
   /// This is the lifecycle state change hook; every time the game is resumed, paused or suspended, this is called.
   ///
@@ -47,7 +48,11 @@ abstract class Game {
   void lifecycleStateChange(AppLifecycleState state) {}
 
   /// Used for debugging
+  @Deprecated('Gets called for backward compatibility, will be removed in v1')
   void recordDt(double dt) {}
+
+  /// Use for caluclating the FPS.
+  void onTimingsCallback(List<FrameTiming> timings) {}
 
   /// Returns the game widget. Put this in your structure to start rendering and updating the game.
   /// You can add it directly to the runApp method or inside your widget structure (if you use vanilla screens and widgets).