+++ lib/sprite.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 
-import 'position.dart';
+import 'extensions/vector2.dart';
 import 'palette.dart';
 
 class Sprite {
@@ -24,30 +24,40 @@ class Sprite {
 
   double get _imageHeight => image.height.toDouble();
 
-  Position get originalSize => Position(_imageWidth, _imageHeight);
+  Vector2 get originalSize => Vector2(_imageWidth, _imageHeight);
 
-  Position get size => Position(src.width, src.height);
+  Vector2 get size => Vector2(src.width, src.height);
 
   /// Renders this Sprite on the position [p], scaled by the [scale] factor provided.
   ///
   /// It renders with src size multiplied by [scale] in both directions.
   /// Anchor is on top left as default.
   /// If not loaded, does nothing.
-  void renderScaled(Canvas canvas, Position p,
-      {double scale = 1.0, Paint overridePaint}) {
-    renderPosition(canvas, p,
-        size: size.times(scale), overridePaint: overridePaint);
+  void renderScaled(
+    Canvas canvas,
+    Vector2 p, {
+    double scale = 1.0,
+    Paint overridePaint,
+  }) {
+    renderPosition(canvas, p, size: size * scale, overridePaint: overridePaint);
   }
 
-  void renderPosition(Canvas canvas, Position p,
-      {Position size, Paint overridePaint}) {
+  void renderPosition(
+    Canvas canvas,
+    Vector2 p, {
+    Vector2 size,
+    Paint overridePaint,
+  }) {
     size ??= this.size;
-    renderRect(canvas, Position.rectFrom(p, size),
-        overridePaint: overridePaint);
+    renderRect(canvas, p.toRect(size), overridePaint: overridePaint);
   }
 
-  void render(Canvas canvas,
-      {double width, double height, Paint overridePaint}) {
+  void render(
+    Canvas canvas, {
+    double width,
+    double height,
+    Paint overridePaint,
+  }) {
     width ??= size.x;
     height ??= size.y;
     renderRect(canvas, Rect.fromLTWH(0.0, 0.0, width, height),
@@ -58,15 +68,23 @@ class Sprite {
   ///
   /// If [size] is not provided, the original size of the src image is used.
   /// If the asset is not yet loaded, it does nothing.
-  void renderCentered(Canvas canvas, Position p,
-      {Position size, Paint overridePaint}) {
+  void renderCentered(
+    Canvas canvas,
+    Vector2 p, {
+    Vector2 size,
+    Paint overridePaint,
+  }) {
     size ??= this.size;
     renderRect(canvas,
         Rect.fromLTWH(p.x - size.x / 2, p.y - size.y / 2, size.x, size.y),
         overridePaint: overridePaint);
   }
 
-  void renderRect(Canvas canvas, Rect dst, {Paint overridePaint}) {
+  void renderRect(
+    Canvas canvas,
+    Rect dst, {
+    Paint overridePaint,
+  }) {
     canvas.drawImageRect(image, src, dst, overridePaint ?? paint);
   }
 }