+++ lib/widgets/animation_widget.dart
@@ -1,10 +1,12 @@
-import 'package:flutter/material.dart';
-import 'package:flame/sprite_animation.dart';
-
 import 'dart:math';
 
+import 'package:flutter/material.dart' hide Animation;
+import 'package:flame/sprite_animation.dart';
+
 import '../anchor.dart';
-import './sprite_widget.dart';
+import '../sprite_animation.dart';
+
+import 'sprite_widget.dart';
 
 /// A [StatefulWidget] that render a [SpriteAnimation].
 class SpriteAnimationWidget extends StatefulWidget {