+++ lib/widgets/sprite_widget.dart
@@ -1,10 +1,11 @@
-import 'package:flame/widgets/animation_widget.dart';
-import 'package:flutter/widgets.dart';
-import 'package:meta/meta.dart';
 import 'dart:math';
 
-import '../sprite.dart';
+import 'package:flutter/widgets.dart';
+import 'package:meta/meta.dart';
+
 import '../anchor.dart';
+import '../sprite.dart';
+import 'animation_widget.dart';
 
 /// A [StatefulWidget] that renders a still [Sprite].
 ///
@@ -50,12 +51,12 @@ class _SpritePainter extends CustomPainter {
     final w = _sprite.size.x * rate;
     final h = _sprite.size.y * rate;
 
-    final double dx = _anchor.relativePosition.dx * size.width;
-    final double dy = _anchor.relativePosition.dy * size.height;
+    final double dx = _anchor.relativePosition.x * size.width;
+    final double dy = _anchor.relativePosition.y * size.height;
 
     canvas.translate(
-      dx - w * _anchor.relativePosition.dx,
-      dy - h * _anchor.relativePosition.dy,
+      dx - w * _anchor.relativePosition.x,
+      dy - h * _anchor.relativePosition.y,
     );
 
     _sprite.render(canvas, width: w, height: h);