+++ test/components/resizable_test.dart
@@ -1,6 +1,5 @@
-import 'dart:ui';
-
 import 'package:flame/game/base_game.dart';
+import 'package:flame/extensions/vector2.dart';
 import 'package:test/test.dart';
 
 import 'package:flame/components/position_component.dart';
@@ -18,7 +17,7 @@ class MyComponent extends PositionComponent with Resizable {
 
 class MyGame extends BaseGame {}
 
-Size size = const Size(1.0, 1.0);
+Vector2 size = Vector2(1.0, 1.0);
 
 void main() {
   group('resizable test', () {